Refactor role to use Docker Compose
Refactors the Sal Docker role to use Docker Compose rather than pure
Ansible for managing containers. In testing, this appeared to more
reliably recreate the interdependant multi-container setup each time it
was run. This change should provide more stability and indempotency.

In the future, some consideration should be given to how to run multiple
compose files on the same machine.